Removed imperative logger
